refactor: extract keybinding tables into tables.ts
